requirements.txt
asgiref==3.8.1
click==8.2.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
gunicorn==23.0.0
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pgvector==0.4.1
pillow==11.2.1
psycopg2==2.9.10
PyJWT==2.9.0
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
wheel==0.45.1
